The WLR S4 Class

The WLR object with the slot information listed as follow:

Slots

project_name

Name of the project.

raw_data

Raw data. A d x M matrix with d rows of features and M columns of data points.

normalized_data

Normalized data. Same shape as raw_data.

scaled_data

Scaled data (z-score). Same shape as raw_data.

pca

Principal component analysis result, see irlba.

dist_null

Null nearest neighbor M x M distance matrix.

sketch_id

Geomertric sketching cell IDs.

sketch_dist

Geomertric sketching distance matrix.

weight_mat

List of feature and sample weight matrice:

feature_weight, n x d binary matrix with n rows of bootstrap iterations and d columns of features where 0 means feature not sampled and 1 means sampled.

sample_weight, n x M matrix with n rows of bootstrap iterations and M columns of data points where each entry represents weight.

perturb_mat, d x M matrix with d rows of features and M columns of data points where each entry represents the relative importance of a feature to a data point.

dist_mat

Permuted distance matrix.

params

Commands used.